Convert to static imports, remove circular deps
